requirements.txt
-c constraints.txt
django
djangorestframework
pyyaml
uritemplate
